get_feedback endpoint and page for rendering it
