+++ backend/src/app.py
@@ -1,9 +1,10 @@
-import os
 from dotenv import load_dotenv
 from flask import Flask, request, jsonify
 
-from chat import chat, CACHE
+from chat import chat
+from shared import CACHE
 from submit_feedback import submit_feedback
+from get_feedback import get_feedback
 load_dotenv(override=True)
 
 app = Flask(__name__)
@@ -13,6 +14,7 @@ def history(session_id):
     return jsonify(CACHE.get(session_id, []))
 
 app.add_url_rule("/api/query", view_func=chat, methods=["POST"])
+app.add_url_rule("/api/get_feedback", view_func=get_feedback, methods=["POST"])
 app.add_url_rule("/api/feedback", view_func=submit_feedback, methods=["POST"])
 
 if __name__ == "__main__":
