+++ backend/src/shared.py
@@ -1,0 +1,4 @@
+from collections import defaultdict
+
+CACHE = defaultdict(list)
+FEEDBACK_FILE = "feedback.jsonl"
